+++ src/main/java/com/opentable/server/ServerConfigConfiguration.java
@@ -11,11 +11,13 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.PropertyResolver;
 
 import com.opentable.service.AppInfo;
 import com.opentable.service.EnvInfo;
 import com.opentable.spring.ConversionServiceConfiguration;
 import com.opentable.spring.PropertySourceUtil;
+import com.opentable.spring.SpecializedConfigFactory;
 
 @Configuration
 @Import({
@@ -39,5 +41,10 @@ public class ServerConfigConfiguration {
                           .sorted()
                           .map(k -> k + "=" + env.getProperty(k))
                           .collect(Collectors.joining("\n" + INDENT)));
+    }
+
+    @Bean
+    public SpecializedConfigFactory<ServerConnectorConfig> connectorConfigs(PropertyResolver pr) {
+        return SpecializedConfigFactory.create(pr, ServerConnectorConfig.class, "ot.httpserver.connector.${name}");
     }
 }