+++ src/main/java/com/opentable/server/ServerConnectorConfig.java
@@ -1,0 +1,23 @@
+package com.opentable.server;
+
+public interface ServerConnectorConfig {
+    default String getProtocol() {
+        return "http";
+    }
+
+    default int getPort() {
+        return -1;
+    }
+
+    default boolean isForceSecure() {
+        return false;
+    }
+
+    default String getKeystore() {
+        return null;
+    }
+
+    default String getKeystorePassword() {
+        return "changeit";
+    }
+}