+++ src/test/java/com/opentable/server/BasicTest.java
@@ -19,6 +19,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -32,6 +34,7 @@ import org.springframework.test.context.junit4.SpringRunner;
     "ot.httpserver.max-threads=13",
     "ot.httpserver.static-path=static-test",
 })
+@DirtiesContext(classMode=ClassMode.AFTER_EACH_TEST_METHOD)
 public class BasicTest {
     @Inject
     LoopbackRequest request;